FEAT/ui

+++ src/app/exercise/sql/select_sets_for_graph_by_exercise_id.sql
@@ -6,4 +6,4 @@ from sets s
 join exercises e on e.id = s.exercise_id
 where s.exercise_id = $1
   and e.user_id = $2
-order by s.created_at::date asc, (s.weight_in_g * (1 + s.repetitions / 30)) desc
+order by s.created_at::date asc, (s.weight_in_g::float * (1.0 + s.repetitions::float / 30.0)) desc
